server/requirements.txt
anyio==3.7.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
fastapi==0.97.0
h11==0.14.0
idna==3.4
pydantic==1.10.9
python-dotenv==1.0.0
redis==4.5.5
requests==2.31.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.27.0
typing_extensions==4.6.3
urllib3==2.0.3
uvicorn==0.22.0
